requirements.txt
requests==2.32.2
python-dotenv==1.1.0